attempting to refactor uploads so they continue after activity os recreated
