+++ app/src/main/AndroidManifest.xml
@@ -28,6 +28,8 @@
     <uses-permission android:name="android.permission.READ_SYNC_SETTINGS"/>
     <!-- Enable sync if it's disabled -->
     <uses-permission android:name="android.permission.WRITE_SYNC_SETTINGS"/>
+    <!-- To be able to uload files in background. See UploadService. -->
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
 
     <application
         android:name=".TindroidApp"
@@ -128,7 +130,7 @@
             android:theme="@style/LicensesTheme" />
 
         <service
-            android:name=".account.TinodeAccountService"
+            android:name=".services.TinodeAccountService"
             android:exported="false">
 
             <!-- Required filter used by the system to launch the account service. -->
@@ -142,7 +144,7 @@
         </service>
 
         <service
-            android:name=".account.SyncService"
+            android:name=".services.ContactsSyncService"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.content.SyncAdapter"/>
@@ -159,12 +161,16 @@
         <!-- FirebaseMessagingService performs security checks at runtime,
              no need for explicit permissions despite exported="true".
              Using tools:ignore="ExportedService" to disable the warning -->
-        <service android:name=".fcm.FBaseMessagingService"
+        <service android:name=".services.FBaseMessagingService"
                  tools:ignore="ExportedService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
         </service>
+
+        <service android:name=".services.UploadService"
+            android:label="Attachment upload service"
+            android:permission="android.permission.BIND_JOB_SERVICE" />
 
         <provider
             android:name="androidx.core.content.FileProvider"
