+++ app/src/main/java/co/tinode/tindroid/services/ContactsSyncService.java
@@ -1,4 +1,4 @@
-package co.tinode.tindroid.account;
+package co.tinode.tindroid.services;
 
 import android.annotation.SuppressLint;
 import android.app.Service;
@@ -15,10 +15,10 @@ import android.os.IBinder;
  * For performance, only one sync adapter will be initialized within this application's context.
  *
  * Note: The SyncService itself is not notified when a new sync occurs. It's role is to
- * manage the lifecycle of our {@link SyncAdapter} and provide a handle to said SyncAdapter to the
+ * manage the lifecycle of our {@link ContactsSyncAdapter} and provide a handle to said SyncAdapter to the
  * OS on request.
  */
-public class SyncService extends Service {
+public class ContactsSyncService extends Service {
     private static final String TAG = "SyncService";
 
     private static final Object sSyncAdapterLock = new Object();
@@ -26,10 +26,10 @@ public class SyncService extends Service {
     // Verbatim copy of
     // https://developer.android.com/training/sync-adapters/creating-sync-adapter
     @SuppressLint("StaticFieldLeak")
-    private static SyncAdapter sSyncAdapter = null;
+    private static ContactsSyncAdapter sSyncAdapter = null;
 
     /**
-     * Thread-safe constructor, creates static {@link SyncAdapter} instance.
+     * Thread-safe constructor, creates static {@link ContactsSyncAdapter} instance.
      */
     @Override
     public void onCreate() {
@@ -37,18 +37,18 @@ public class SyncService extends Service {
 
         synchronized (sSyncAdapterLock) {
             if (sSyncAdapter == null) {
-                sSyncAdapter = new SyncAdapter(getApplicationContext(), true);
+                sSyncAdapter = new ContactsSyncAdapter(getApplicationContext(), true);
             }
         }
     }
 
     /**
-     * Return Binder handle for IPC communication with {@link SyncAdapter}.
+     * Return Binder handle for IPC communication with {@link ContactsSyncAdapter}.
      * <p>
      * <p>New sync requests will be sent directly to the SyncAdapter using this channel.
      *
      * @param intent Calling intent
-     * @return Binder handle for {@link SyncAdapter}
+     * @return Binder handle for {@link ContactsSyncAdapter}
      */
     @Override
     public IBinder onBind(Intent intent) {
