+++ build.gradle
@@ -8,7 +8,7 @@ buildscript {
         maven { url 'https://maven.fabric.io/public' }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.2'
+        classpath 'com.android.tools.build:gradle:3.6.3'
 
         // For FCM push notifications
         classpath 'com.google.gms:google-services:4.3.3'
